Add the thread0 package.
This package contains the code for scheduling functions to execute on
teh main go routine. It both reads more clearly than ui.Do (i.e.,
thread0.Do makes it obvious that the thing being done is by the 0th
thread, not somehow by the ui.) and removes a dependency on the rest
of the ui code for things that simply want to execute on the 0th
thread.